+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add buttons for navigation on the Connected page ([#722])
 * Improve tooltip behavior ([#723])
 * Better settings controls ([#725])
+* Rework patch visualizer with many fixes and improvements ([#726])
 * Added better support for `Font` properties ([#731])
 
 [#668]: https://github.com/rojo-rbx/rojo/pull/668
@@ -31,6 +32,7 @@
 [#722]: https://github.com/rojo-rbx/rojo/pull/722
 [#723]: https://github.com/rojo-rbx/rojo/pull/723
 [#725]: https://github.com/rojo-rbx/rojo/pull/725
+[#726]: https://github.com/rojo-rbx/rojo/pull/726
 [#731]: https://github.com/rojo-rbx/rojo/pull/731
 
 ## [7.3.0] - April 22, 2023